refactor: align tools naming
